@@ -78,49 +78,51 @@
     ]
   },
   "dependencies": {
-    "@angular/core": "8.0.0",
-    "@nrwl/angular": "8.0.0",
-    "core-js": "2.5.7",
+    "@angular/core": "8.1.1",
+    "@nrwl/angular": "8.2.0",
+    "@types/fontfaceobserver": "^0.0.6",
+    "core-js": "2.6.9",
+    "fontfaceobserver": "^2.1.0",
     "ij-rpc-client": "^0.3.2",
     "node-pty": "^0.8.1",
+    "pretty-quick": "^1.11.1",
     "source-map-explorer": "^1.8.0",
-    "stylelint": "^9.10.1",
-    "stylelint-config-standard": "^18.2.0",
-    "tmp": "^0.1.0"
+    "stylelint": "^10.1.0",
+    "stylelint-config-standard": "^18.3.0",
+    "tmp": "^0.1.0",
+    "webpack": "^4.35.3"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "0.800.0",
     "@angular-devkit/build-ng-packagr": "0.800.0",
-    "@angular/animations": "8.0.0",
-    "@angular/cdk": "8.0.0",
-    "@angular/cli": "8.0.0",
-    "@angular/common": "8.0.0",
-    "@angular/compiler": "8.0.0",
-    "@angular/compiler-cli": "8.0.0",
+    "@angular/animations": "8.1.1",
+    "@angular/cdk": "8.1.0",
+    "@angular/cli": "8.1.1",
+    "@angular/common": "8.1.1",
+    "@angular/compiler": "8.1.1",
+    "@angular/compiler-cli": "8.1.1",
     "@angular/flex-layout": "8.0.0-beta.26",
-    "@angular/forms": "8.0.0",
-    "@angular/language-service": "8.0.0",
-    "@angular/material": "8.0.0",
-    "@angular/platform-browser": "8.0.0",
-    "@angular/platform-browser-dynamic": "8.0.0",
-    "@angular/router": "8.0.0",
+    "@angular/forms": "8.1.1",
+    "@angular/language-service": "8.1.1",
+    "@angular/material": "8.1.0",
+    "@angular/platform-browser": "8.1.1",
+    "@angular/platform-browser-dynamic": "8.1.1",
+    "@angular/router": "8.1.1",
     "@graphql-modules/logger": "^0.2.3",
     "@nestjs/common": "5.7.0",
     "@nestjs/core": "5.7.0",
     "@nestjs/graphql": "5.5.1",
     "@nrwl/angular-console-enterprise-electron": "^0.3.681",
     "@nrwl/angular-console-enterprise-frontend": "^0.3.681",
-    "@nrwl/cypress": "8.0.0",
-    "@nrwl/jest": "8.0.0",
-    "@nrwl/node": "8.0.0",
-    "@nrwl/workspace": "8.0.0",
+    "@nrwl/cypress": "8.2.0",
+    "@nrwl/jest": "8.2.0",
+    "@nrwl/node": "8.2.0",
+    "@nrwl/workspace": "8.2.0",
     "@types/d3-hierarchy": "^1.1.6",
     "@types/d3-selection": "^1.4.1",
     "@types/d3-shape": "^1.3.1",
     "@types/d3-transition": "^1.1.4",
     "@types/electron-store": "^3.2.0",
-    "@types/fontfaceobserver": "^0.0.6",
-    "@types/get-port": "^4.0.1",
     "@types/graphql": "14.0.5",
     "@types/jest": "24.0.9",
     "@types/node": "11.13.0",
@@ -141,7 +143,7 @@
     "apollo-link-error": "1.1.11",
     "apollo-server-express": "^2.6.9",
     "codelyzer": "5.1.0",
-    "cypress": "^3.1.5",
+    "cypress": "3.4.0",
     "d3-hierarchy": "^1.1.8",
     "d3-selection": "^1.4.0",
     "d3-shape": "^1.3.4",
@@ -155,10 +157,9 @@
     "electron-store": "3.3.0",
     "electron-updater": "4.1.2",
     "express": "4.16.4",
-    "fast-glob": "^2.2.6",
+    "fast-glob": "3.0.4",
     "fix-path": "2.1.0",
-    "fontfaceobserver": "^2.1.0",
-    "get-port": "4.1.0",
+    "get-port": "5.0.0",
     "graphql": "14.2.1",
     "graphql-code-generator": "0.16.1",
     "graphql-codegen-typescript-apollo-angular": "0.16.1",
@@ -227,7 +228,7 @@
   },
   "husky": {
     "hooks": {
-      "pre-commit": "yarn start format && git add ."
+      "pre-commit": "pretty-quick --staged && git add ."
     }
   }
 }
